@@ -1,31 +1,31 @@
 <template>
   <div class="event-page">
 
-    <button class="back-button" @click="$router.go(-1)">←</button>
+    <button class="back-button" @click="router.go(-1)">←</button>
 
     <div class="event-left">
-      <img class="event-image" :src="eventData.image" alt="Event Image" />
+      <img class="event-image" :src="eventData?.image" alt="Event Image" />
 
       <div class="description">
         <h3>Description</h3>
-        <p>{{ eventData.description }}</p>
+        <p>{{ eventData?.description }}</p>
       </div>
     </div>
 
     <div class="event-right">
-      <h1>{{ eventData.name }}</h1>
+      <h1>{{ eventData?.name }}</h1>
       <button class="register-button" @click="goToTicketSelection">Register Now</button>
 
       <div class="event-info">
-        <p><strong>When</strong><br>{{ eventData.date }}</p>
-        <p><strong>Location</strong><br>{{ eventData.location }}</p>
-        <p><strong>Visit organizer's website:</strong> <a :href="eventData.website" target="_blank">{{ eventData.website }}</a></p>
+        <p><strong>When</strong><br>{{ eventData?.date }}</p>
+        <p><strong>Location</strong><br>{{ eventData?.location }}</p>
+        <p><strong>Visit organizer's website:</strong> <a :href="eventData?.website" target="_blank">{{ eventData?.website }}</a></p>
       </div>
 
       <!-- Using the part of the map that comes with VSCode -->
       <div class="event-map">
         <iframe
-          :src="eventData.mapUrl"
+          :src="eventData?.mapUrl"
           width="100%"
           height="200"
           style="border:0;"
@@ -36,60 +36,58 @@
 
       <div class="share-link">
         <p><strong>Share:</strong></p>
-        <input type="text" readonly :value="eventData.shareLink" />
+        <input type="text" readonly :value="eventData?.shareLink" />
       </div>
     </div>
   </div>
 </template>
 
-<script>
-export default {
-  data() {
-    return {
-      events: [
-        {
-          id: 1,
-          name: 'Melbourne Cricket Ground Event',
-          description:
-            'Lorem ipsum odor amet, consectetuer adipiscing elit. Nunc sociosqu a hac semper nullam lectus laoreet. Vestibulum aenean condimentum ornare a ex ridiculus lacinia tellus per.',
-          date: 'October 13, 2024, Sunday',
-          location: 'Melbourne Cricket Ground (MCG), Hawthorn',
-          website: 'https://example.com',
-          // VSCode Maps Related Links
-          mapUrl: 'https://maps.google.com/maps?q=Melbourne%20Cricket%20Ground&t=&z=13&ie=UTF8&iwloc=&output=embed',
-          image: require('@/assets/Melbourne-Cricket-Ground.png'), 
-          shareLink: 'https://link.to.event1.com',
-        },
-        {
-          id: 2,
-          name: 'Event 2',
-          description: 'Detailed description for Event 2.',
-          date: 'October 14, 2024, Monday',
-          location: 'Sydney Opera House, Sydney',
-          mapUrl: 'https://maps.google.com/maps?q=Sydney%20Opera%20House&t=&z=13&ie=UTF8&iwloc=&output=embed',
-          image: 'https://via.placeholder.com/600x400',
-          shareLink: 'https://link.to.event2.com',
-        },
-        // More events...
-      ],
-      eventData: null,
-    };
+<script setup>
+import { ref, onMounted } from 'vue'
+import { useRouter, useRoute } from 'vue-router'
+
+const router = useRouter()
+const route = useRoute()
+
+const events = ref([
+  {
+    id: 1,
+    name: 'Melbourne Cricket Ground Event',
+    description:
+      'Lorem ipsum odor amet, consectetuer adipiscing elit. Nunc sociosqu a hac semper nullam lectus laoreet. Vestibulum aenean condimentum ornare a ex ridiculus lacinia tellus per.',
+    date: 'October 13, 2024, Sunday',
+    location: 'Melbourne Cricket Ground (MCG), Hawthorn',
+    website: 'https://example.com',
+    mapUrl: 'https://maps.google.com/maps?q=Melbourne%20Cricket%20Ground&t=&z=13&ie=UTF8&iwloc=&output=embed',
+    image: new URL('../assets/images/Melbourne-Cricket-Ground.png', import.meta.url).href,
+    shareLink: 'https://link.to.event1.com',
   },
-  created() {
-    const eventId = this.$route.params.id;
-    this.loadEventDetails(eventId);
+  {
+    id: 2,
+    name: 'Event 2',
+    description: 'Detailed description for Event 2.',
+    date: 'October 14, 2024, Monday',
+    location: 'Sydney Opera House, Sydney',
+    mapUrl: 'https://maps.google.com/maps?q=Sydney%20Opera%20House&t=&z=13&ie=UTF8&iwloc=&output=embed',
+    image: new URL('../assets/images/Melbourne-Cricket-Ground.png', import.meta.url).href,
+    shareLink: 'https://link.to.event2.com',
   },
-  methods: {
-    loadEventDetails(id) {
-      // Find matching event data
-      this.eventData = this.events.find(event => event.id === parseInt(id, 10));
-    },
-    goToTicketSelection() {
-      // Skip to ticket selection page
-      this.$router.push(`/ticket-selection`);
-    }
-  }
-};
+])
+
+const eventData = ref(null)
+
+const loadEventDetails = (id) => {
+  eventData.value = events.value.find(event => event.id === parseInt(id, 10))
+}
+
+const goToTicketSelection = () => {
+  router.push('/ticket-selection')
+}
+
+onMounted(() => {
+  const eventId = route.params.id
+  loadEventDetails(eventId)
+})
 </script>
 
 <style scoped>
