@@ -1,6 +1,13 @@
-import { createApp } from "vue";
-import App from "./App.vue";
-import router from "./router";
+import "@/assets/css/main.css"
+
+
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
 import 'leaflet/dist/leaflet.css';
 
-createApp(App).use(router).mount("#app");
+const app = createApp(App)
+
+app.use(router)
+
+app.mount('#app')
